refactor: simplify tests
@NilsJacobsen @NiklasBuchfink the tests were likely failing due to custom tests that used hand picked ASTs instead of using readFile.

It was a mess to go through the tests. I cleaned up things and hope that I didn't remove something important.

A few improvements for next time:

- use multiple tests. Don't put everything in one test.
- use one test file that goes through every example
- don't hardcode ASTs. Use the ones that are in the examples itself through readFile
