--- example09/en/translation.json
@@ -1,3 +1,0 @@
-{
-  "message01": "first"
-}
