--- exampleTry/en/test.json
@@ -1,3 +1,0 @@
-{
-  "test": "asdfas"
-}
