+++ examples/example04/inlang.config.js
@@ -1,0 +1,16 @@
+/**
+ * @type {import("@inlang/core/config").DefineConfig}
+ */
+export async function defineConfig(env) {
+  const { default: plugin } = await env.$import("./dist/index.js");
+
+  return {
+    referenceLanguage: "en",
+    plugins: [
+      plugin({
+        pathPattern: "./examples/example04/{language}.json",
+        variableReferencePattern: ["{", "}"],
+      }),
+    ],
+  };
+}
