+++ examples/example05/fr/translation.json
