+++ examples/example06/en/translation.json
